entire_codebase/requirements/training.txt
apscheduler==3.10.4
mlflow==3.1.0
numpy==2.0.2
pandas==2.3.0
PyYAML==6.0.2
ray[train]==2.51.2
scipy==1.13.1
xgboost==2.1.4
prometheus-client>=0.20.0
